% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_islay_land.R
\docType{data}
\name{islay_land}
\alias{islay_land}
\title{Island of Islay}
\format{
A simple feature (\link[sf:sf]{sf::sf}) object with 255 features and 2 fields:
\describe{
\item{name}{Name of the island, if available.}
\item{area}{Area of the island in square metres.}
}
}
\usage{
islay_land
}
\description{
Spatial geometries for the islands Islay, Texa, Nave Island, Orsay, and
251 associated islets.
}
\references{
\itemize{
\item Derived from \href{https://www.openstreetmap.org/copyright}{OpenStreetMap} data
}
}
\keyword{datasets}
