% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_islay_lithics.R
\docType{data}
\name{islay_lithics}
\alias{islay_lithics}
\title{Lithic assemblages from Islay}
\format{
A data frame with 75 rows and 10 variables:
\describe{
\item{site_code}{Character. Abbreviated name of the site. Corresponds to the \code{site_code} column of \link{islay_sites}.}
\item{region}{Character. Regional grouping the site belongs to. One of: \code{"Loch Gorm South"}, \code{"Loch Gorm North"}, \code{"Bowmore Coastal"}, \code{"Bowmore Inland"}, \code{"Bridgend"}, \code{"Kiells"}, \code{"Loch Gruinart"}, or \code{"Port Charlotte"}.}
\item{period}{Character. Period assigned to the site by expert analysis. One of: \code{"Mesolithic"}, \code{"Mesolithic & Later Prehistoric"}, \code{"Later Prehistoric"}, or \code{NA}.}
\item{area}{Numeric. Area of the site in square meters.}
\item{flakes,blades,chunks,cores,pebbles,retouched}{Numeric. Frequency of lithic components recorded at the site.}
\item{total}{Numeric. Total number of lithics recoverd from the site (sum of \code{flakes:retouched}).}
}
}
\usage{
islay_lithics
}
\description{
This dataset describes 75 chipped stone assemblages from surface sites on the
Scottish island of Islay, recorded by the Southern Hebrides Mesolithic
Project (SHMP).
}
\references{
\insertRef{SHMP4.2}{islay}
}
\keyword{datasets}
